Added merging
