+++ MyApp/MyApp.Core/MyApp.Core.csproj
@@ -10,6 +10,7 @@
     <PackageReference Include="Elasticsearch.Net" Version="7.17.5" />
     <PackageReference Include="Microsoft.Extensions.Caching.StackExchangeRedis" Version="9.0.2" />
     <PackageReference Include="NEST" Version="7.17.5" />
+    <PackageReference Include="npgsql" Version="9.0.3" />
     <PackageReference Include="RabbitMQ.Client" Version="6.5.0" />
     <PackageReference Include="StackExchange.Redis" Version="2.8.31" />
     <PackageReference Include="Swashbuckle.AspNetCore" Version="7.3.1" />
