+++ MyApp/MyApp.Core/Repositories/Interfaces/IAdminInterface.cs
@@ -7,6 +7,10 @@ namespace MyApp.Core.Repositories.Interfaces
 {
     public interface IAdminInterface
     {
-        
+        Task<List<t_user>> GetAll();
+         Task<List<t_task>> GetTaskDocument();
+          int GetTotalUsers();
+           int GetTotalTasks();
+               Task<(int totalUsers, int totalTasks)> GetDashboardStatsAsync();
     }
 }
