+++ MyApp/MyApp.Core/Services/RabbitMQService.cs
@@ -5,81 +5,39 @@ using System.Text;
 using System.Threading.Tasks;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.Hosting;
+using Npgsql;
 using RabbitMQ.Client;
 using RabbitMQ.Client.Events;
 
+
 namespace MyApp.Core.Services
 {
-    public class RabbitMQService : BackgroundService
+    public class RabbitMQService
     {
-        private readonly string _queueName = "UserRegistrationQueue";
-        private readonly ConnectionFactory _factory;
-        private IConnection _connection;
-        private IModel _channel;
-        private readonly string _connectionString;
+        private readonly string _hostName;
+        private readonly string _userName;
+        private readonly string _password;
+        private readonly string _queueName;
 
         public RabbitMQService(IConfiguration configuration)
         {
-            _factory = new ConnectionFactory()
-            {
-                HostName = "localhost",
-                UserName = "guest",
-                Password = "guest"
-            };
-
-            _connectionString = configuration.GetConnectionString("pgconn");
+            _hostName = configuration["RabbitMQ:HostName"];
+            _userName = configuration["RabbitMQ:UserName"];
+            _password = configuration["RabbitMQ:Password"];
+            _queueName = configuration["RabbitMQ:QueueName"];
         }
-        protected override Task ExecuteAsync(CancellationToken stoppingToken)
+        public void PublishMessage(string message)
         {
-            _connection = _factory.CreateConnection();
-            _channel = _connection.CreateModel();
+            var factory = new ConnectionFactory() { HostName = _hostName, UserName = _userName, Password = _password };
+            using var connection = factory.CreateConnection();
+            using var channel = connection.CreateModel();
 
-            _channel.QueueDeclare(queue: _queueName, durable: false, exclusive: false, autoDelete: false, arguments: null);
+            channel.QueueDeclare(queue: _queueName, durable: false, exclusive: false, autoDelete: false, arguments: null);
 
-            var consumer = new EventingBasicConsumer(_channel);
-            consumer.Received += (model, ea) =>
-            {
-                var body = ea.Body.ToArray();
-                var message = Encoding.UTF8.GetString(body);
+            var body = Encoding.UTF8.GetBytes(message);
 
-                Console.WriteLine($"[Notification Received]: {message}");
-
-                // Save the notification into PostgreSQL
-                SaveNotificationToDatabase(message);
-            };
-
-            _channel.BasicConsume(queue: _queueName, autoAck: true, consumer: consumer);
-
-            return Task.CompletedTask;
+            channel.BasicPublish(exchange: "", routingKey: _queueName, basicProperties: null, body: body);
         }
-
-        private void SaveNotificationToDatabase(string message)
-        {
-            try
-            {
-                using var conn = new NpgsqlConnection(_connectionString);
-                conn.Open();
-
-                string sql = "INSERT INTO t_notification (c_title, c_taskId, c_userId) VALUES (@title, NULL, NULL)";
-
-                using var cmd = new NpgsqlCommand(sql, conn);
-                cmd.Parameters.AddWithValue("@title", message);
-
-                cmd.ExecuteNonQuery();
-            }
-            catch (Exception ex)
-            {
-                Console.WriteLine($"[Database Error]: {ex.Message}");
-            }
-        }
-
-        public override void Dispose()
-        {
-            _channel?.Close();
-            _connection?.Close();
-            base.Dispose();
-        }
-
 
     }
 }
